backend/requirements.txt
Flask==3.1.0
flask-cors==4.0.0
yt-dlp>=2024.03.10
python-dotenv==1.0.0
requests==2.31.0
